fix(chat): Changes requested in code review
