+++ bigbluebutton-html5/imports/ui/components/chat/chat-graphql/chat-message-list/component.tsx
@@ -202,8 +202,8 @@ const ChatMessageList: React.FC<ChatListProps> = ({
   const [followingTail, setFollowingTail] = React.useState(true);
   const [showStartSentinel, setShowStartSentinel] = React.useState(false);
   const [focusedMessageElement, setFocusedMessageElement] = React.useState<HTMLElement | null>();
-  const [loadingPages, setLoadingPages] = React.useState(new Set<number>());
-  const allPagesLoaded = loadingPages.size === 0;
+  const [loadingPages, setLoadingPages] = useState(new Set<number>());
+  const [lockLoadingNewPages, setLockLoadingNewPages] = useState(true);
   const {
     childRefProxy: endSentinelRefProxy,
     intersecting: isEndSentinelVisible,
@@ -304,9 +304,10 @@ const ChatMessageList: React.FC<ChatListProps> = ({
       if (followingTail) {
         toggleFollowingTail(false);
       }
-      if (!allPagesLoaded) return;
+      if (lockLoadingNewPages) return;
       setUserLoadedBackUntilPage((prev) => {
         if (typeof prev === 'number' && prev > 0) {
+          setLockLoadingNewPages(true);
           return prev - 1;
         }
         return prev;
@@ -474,6 +475,10 @@ const ChatMessageList: React.FC<ChatListProps> = ({
     });
   }, [setLoadingPages]);
 
+  useEffect(() => {
+    setLockLoadingNewPages(loadingPages.size !== 0);
+  }, [loadingPages]);
+
   return (
     <>
       {
@@ -494,9 +499,10 @@ const ChatMessageList: React.FC<ChatListProps> = ({
               }
             }}
             onWheel={(e) => {
-              if (e.deltaY < 0 && isStartSentinelVisible && allPagesLoaded) {
+              if (e.deltaY < 0 && isStartSentinelVisible && !lockLoadingNewPages) {
                 setUserLoadedBackUntilPage((prev) => {
                   if (typeof prev === 'number' && prev > 0) {
+                    setLockLoadingNewPages(true);
                     return prev - 1;
                   }
                   return prev;
