+++ bigbluebutton-html5/imports/ui/components/chat/chat-graphql/chat-message-list/page/loader/component.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { CircularProgress } from '@mui/material';
+import Styled from './styles';
+
+const ChatPageLoading = () => {
+  return (
+    <Styled.ChatPageLoading>
+      <CircularProgress />
+    </Styled.ChatPageLoading>
+  );
+};
+
+export default ChatPageLoading;
