+++ bigbluebutton-html5/imports/ui/components/chat/chat-graphql/chat-message-list/page/loader/styles.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const ChatPageLoading = styled.div`
+  display: flex;
+  justify-content: center;
+`;
+
+export default {
+  ChatPageLoading,
+};
